fix: split migration finalization into separate phase, fix PVC rebind race
Add FinalizingFilestoreMigration phase to separate the rsync wait
(MigratingFilestore) from the PVC rebind (FinalizingFilestoreMigration).

Key fixes:
- Transition action only stores PV name + deletes Job (fast, can't fail
  on PVC state)
- PVC rebind (delete old, patch PV, create final) runs in ensure() with
  idempotent retries — returns Ok if old PVC still terminating
- Job deleted with foreground propagation before PVC deletion to release
  pvc-protection finalizer
- Absent job in MigratingFilestore → rollback (rsync never completed)
- Absent job in FinalizingFilestoreMigration → impossible (already
  transitioned past rsync)
- Webhook allows rollback (SC change back to migrationPreviousStorageClass)
  even during blocked phases

Also increases backup job activeDeadlineSeconds from 3600 to 5400.
