+++ src/controller/states/mod.rs
@@ -14,10 +14,10 @@ use super::state_machine::ReconcileSnapshot;
 mod backing_up;
 mod degraded;
 mod error;
-mod finalizing_filestore_migration;
+pub(crate) mod finalizing_filestore_migration;
 mod init_failed;
 mod initializing;
-mod migrating_filestore;
+pub(crate) mod migrating_filestore;
 mod provisioning;
 mod restoring;
 mod running;
